Updated the BalanceDisplay component to function more as a stand alone dashboard.
Added cards to display the Net Income/Loss, Total Income, and Total Expenses values

# TODO
Add chart functionality to the charts section. I need to add some mock data to the API so that I can test out these charts with actual data
